Extract Codex OAuth route detection into codex_routing.py (WS5 seam)
Move the four Codex-route helpers and the effective context-cap table out
of engine.py into a dedicated codex_routing.py module: _bare_model_slug,
_is_openai_codex_route, _codex_oauth_context_cap, _is_codex_gpt55_route, and
_CODEX_OAUTH_CONTEXT_CAPS. engine.py imports the two it calls
(_codex_oauth_context_cap, _is_codex_gpt55_route); the internal helpers and
the caps table travel with them. The gpt-5.5 compaction threshold constant
(_CODEX_GPT55_COMPACTION_THRESHOLD) stays in engine.py as engine-side policy.

Behaviour-preserving refactor (WS5 engine decomposition). The moved code is
byte-identical to its pre-move form. Rollback: git revert this commit
restores the inline block; no data/state migration involved.
