+++ engine.py
@@ -23,6 +23,10 @@ from typing import Any, Dict, Iterator, List, Optional
 
 from agent.context_engine import ContextEngine
 
+from .codex_routing import (
+    _codex_oauth_context_cap,
+    _is_codex_gpt55_route,
+)
 from .config import LCMConfig
 from .dag import SummaryDAG, SummaryNode
 from .diagnostics import _enforce_state_db_containment
@@ -186,67 +190,7 @@ def resolve_active_lcm_engine(session_id: str = "", conversation_id: str = "") -
 _PLUGIN_ROOT = Path(__file__).resolve().parent
 _PLUGIN_METADATA: dict[str, str] | None = None
 _SESSION_END_BUSY_TIMEOUT_MS = 50
-# ChatGPT Codex OAuth exposes provider-enforced context windows that can be
-# materially lower than the same model slug on direct OpenAI/OpenRouter routes.
-# Hermes Agent resolves these from chatgpt.com/backend-api/codex/models, with
-# this table as its fallback. LCM sees only the host-advertised context_length;
-# when that value was explicitly overridden above the real Codex OAuth window,
-# we still have to budget against the effective provider window or compaction
-# fires too late and provider requests can overflow.
-_CODEX_OAUTH_CONTEXT_CAPS: dict[str, int] = {
-    "gpt-5.1-codex-max": 272_000,
-    "gpt-5.1-codex-mini": 272_000,
-    "gpt-5.3-codex-spark": 128_000,
-    "gpt-5.3-codex": 272_000,
-    "gpt-5.2-codex": 272_000,
-    "gpt-5.4-mini": 272_000,
-    "gpt-5.5": 272_000,
-    "gpt-5.4": 272_000,
-    "gpt-5.2": 272_000,
-    "gpt-5": 272_000,
-}
 _CODEX_GPT55_COMPACTION_THRESHOLD = 0.85
-
-
-def _bare_model_slug(model: str | None) -> str:
-    return (model or "").strip().lower().rsplit("/", 1)[-1]
-
-
-def _is_openai_codex_route(provider: str | None) -> bool:
-    return (provider or "").strip().lower() == "openai-codex"
-
-
-def _codex_oauth_context_cap(model: str | None, provider: str | None) -> int | None:
-    """Return LCM's best-known Codex OAuth effective context cap.
-
-    This intentionally mirrors Hermes Agent's hardcoded fallback policy, not the
-    direct OpenAI model catalog. A host-provided context_length may be a user
-    override or stale cache entry; Codex OAuth still enforces these lower route
-    windows.
-    """
-    if not _is_openai_codex_route(provider):
-        return None
-    bare_model = _bare_model_slug(model)
-    if not bare_model:
-        return None
-    for slug, cap in sorted(
-        _CODEX_OAUTH_CONTEXT_CAPS.items(), key=lambda item: len(item[0]), reverse=True
-    ):
-        if slug in bare_model:
-            return cap
-    return None
-
-
-def _is_codex_gpt55_route(model: str | None, provider: str | None) -> bool:
-    """Return True for gpt-5.5 on ChatGPT Codex OAuth, mirroring Hermes core."""
-    if not _is_openai_codex_route(provider):
-        return False
-    bare_model = _bare_model_slug(model)
-    return (
-        bare_model == "gpt-5.5"
-        or bare_model.startswith("gpt-5.5-")
-        or bare_model.startswith("gpt-5.5.")
-    )
 
 # Auto-focus topic derivation: infer a compact focus hint from the most recent
 # real user turns so that summarization can prioritise current user intent.
